add filter

+++ Store.Application/Services/Users/Queries/GetUsers/ResultGetUsersDto.cs
@@ -9,6 +9,6 @@ namespace Store.Application.Services.Users.Queries.GetUsers
     public class ResultGetUsersDto
     {
         public List<GetUsersDto> Users { get; set; }
-        public long Rows;
+        public int Rows;
     }
 }
